# Add omp parallel optimization for _contrib_BilinearReisze2D